requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==4.1.2
django-on-heroku==1.1.2
gunicorn==20.1.0
psycopg2-binary==2.9.5
python-decouple==3.6
sqlparse==0.4.3
whitenoise==6.2.0
